+++ demos/py_scripts/dualsphysics.py
@@ -30,7 +30,7 @@ def main(_):
 
     inductiva.api_url = FLAGS.api_url
 
-    sph_sim = inductiva.fluids.simulators.DualSPHysics()
+    sph_sim = inductiva.simulators.DualSPHysics()
 
     task = sph_sim.run(
         input_dir=FLAGS.sim_dir,
