--- inductiva/coastal/simulators/__init__.py
@@ -1,3 +1,0 @@
-#pylint: disable=missing-module-docstring
-from .xbeach import XBeach
-from .swash import SWASH
