+++ inductiva/fluids/__init__.py
@@ -24,8 +24,3 @@ from . import shapes
 from . import post_processing
 from .post_processing import SteadyStateOutput
 from .scenarios._post_processing import SPHSimulationOutput
-from .simulators import (
-    DualSPHysics,
-    SPlisHSPlasH,
-    OpenFOAM,
-)
