--- inductiva/fluids/simulators/__init__.py
@@ -1,4 +1,0 @@
-#pylint: disable=missing-module-docstring
-from .splishsplash import SPlisHSPlasH
-from .dualsphysics import DualSPHysics
-from .openfoam import OpenFOAM
