+++ inductiva/molecules/scenarios/md_water_box/md_water_box.py
@@ -7,7 +7,7 @@ import io
 
 from inductiva import tasks, resources
 from inductiva.molecules.simulators import GROMACS
-from inductiva.simulation import Simulator
+from inductiva.simulators import Simulator
 from inductiva.utils import templates
 from inductiva.scenarios import Scenario
 from .post_processing import MDWaterBoxOutput
