--- inductiva/simulation/__init__.py
@@ -1,2 +1,0 @@
-# pylint: disable=missing-module-docstring
-from .simulator import Simulator
