+++ inductiva/simulators/__init__.py
@@ -1,0 +1,7 @@
+# pylint: disable=missing-module-docstring
+from .simulator import Simulator
+from .swash import SWASH
+from .xbeach import XBeach
+from .splishsplash import SPlisHSPlasH
+from .dualsphysics import DualSPHysics
+from .openfoam import OpenFOAM
