+++ inductiva/simulators/dualsphysics.py
@@ -1,38 +1,35 @@
-"""OpenFOAM module of the API for fluid dynamics."""
-from typing import Optional, List
+"""DualSPHysics module of the API."""
+
+from typing import Literal, Optional
 
 from inductiva import types, tasks, resources
-from inductiva.simulation import Simulator
+from inductiva.simulators import Simulator
 
 
-class OpenFOAM(Simulator):
+class DualSPHysics(Simulator):
     """Class to invoke a generic DualSPHysics simulation on the API."""
-
-    def __init__(self, api_method: str = "fvm"):
-        super().__init__()
-        self.api_method = api_method + ".openfoam.run_simulation"
 
     @property
     def api_method_name(self) -> str:
-        return self.api_method
+        return "sph.dualsphysics.run_simulation"
 
     def run(
         self,
         input_dir: types.Path,
-        commands: List[dict],
+        sim_config_filename: str,
+        device: Literal["gpu", "cpu"] = "cpu",
         machine_group: Optional[resources.MachineGroup] = None,
-        n_cores: int = 1,
         run_async: bool = False,
     ) -> tasks.Task:
         """Run the simulation.
 
         Args:
-            commands: List of commands to run using the OpenFOAM simulator.
-            n_cores: Number of MPI cores to use for the simulation.
+            device: Device in which to run the simulation.
+            sim_config_filename: Name of the simulation configuration file.
             other arguments: See the documentation of the base class.
         """
         return super().run(input_dir,
                            machine_group=machine_group,
-                           n_cores=n_cores,
-                           commands=commands,
+                           device=device,
+                           input_filename=sim_config_filename,
                            run_async=run_async)
