+++ inductiva/simulators/openfoam.py
@@ -1,21 +1,25 @@
-"""SWASH module of the API."""
-from typing import Optional
+"""OpenFOAM module of the API for fluid dynamics."""
+from typing import Optional, List
 
-from inductiva.simulation import Simulator
 from inductiva import types, tasks, resources
+from inductiva.simulators import Simulator
 
 
-class SWASH(Simulator):
-    """Class to invoke a generic SWASH simulation on the API."""
+class OpenFOAM(Simulator):
+    """Class to invoke a generic DualSPHysics simulation on the API."""
+
+    def __init__(self, api_method: str = "fvm"):
+        super().__init__()
+        self.api_method = api_method + ".openfoam.run_simulation"
 
     @property
     def api_method_name(self) -> str:
-        return "sw.swash.run_simulation"
+        return self.api_method
 
     def run(
         self,
         input_dir: types.Path,
-        sim_config_filename: str,
+        commands: List[dict],
         machine_group: Optional[resources.MachineGroup] = None,
         n_cores: int = 1,
         run_async: bool = False,
@@ -23,12 +27,12 @@ class SWASH(Simulator):
         """Run the simulation.
 
         Args:
+            commands: List of commands to run using the OpenFOAM simulator.
             n_cores: Number of MPI cores to use for the simulation.
-            sim_config_filename: Name of the simulation configuration file.
             other arguments: See the documentation of the base class.
         """
         return super().run(input_dir,
                            machine_group=machine_group,
-                           input_filename=sim_config_filename,
                            n_cores=n_cores,
+                           commands=commands,
                            run_async=run_async)
