+++ inductiva/structures/scenarios/plate_linear_elastic/plate_linear_elastic.py
@@ -8,7 +8,7 @@ from functools import singledispatchmethod
 
 from inductiva import tasks, resources
 from inductiva.scenarios import Scenario
-from inductiva.simulation import Simulator
+from inductiva.simulators import Simulator
 from inductiva.structures.simulators import FEniCSx
 from inductiva.types import Path
